Terminal projects/Games/MiniGames/word_game/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
mariadb==1.1.10
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.1.0
urllib3==2.2.2